refactor: create seperate PhotosGrid component and WithLoading component
