+++ src/components/WithLoading/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Loader from '../Loader';
+
+const WithLoading = Component => {
+    return ({ isLoading, ...props }) => isLoading ? <Loader /> : <Component {...props} />;
+};
+
+export default WithLoading;
