+++ src/pages/Photos/index.js
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
-import Card from '../../components/Card';
-import Grid from '../../components/Grid';
 import { PageHeadline, PhotoHeadContainer } from './Photos.styled';
 import Pagination from '../../components/Pagination';
 import PropTypes from 'prop-types';
 import NoResult from '../../components/NoResult';
-import ModalZoom from '../../components/ModalZoom';
 import axios from 'axios';
-import Loader from '../../components/Loader';
 import { checkErrorsFromAPI } from '../../utils';
 import Search from '../../components/Search';
+import PhotosGrid from '../../components/PhotosGrid';
+import WithLoading from '../../components/WithLoading';
+
+const PhotosGridWithLoading = WithLoading(PhotosGrid);
 
 
 const Photos = ({ apiUrl, album }) => {
@@ -19,27 +19,12 @@ const Photos = ({ apiUrl, album }) => {
     const [filteredValue, setFilteredValue] = useState('');
     const [isError, setIsError] = useState(false);
     const [isLoading, setIsLoading] = useState(false);
-    const [clickedCard, setClickedCard] = useState('');
-    const [isModalOpen, setIsModalOpen] = useState(false);
     const [cardsPerPage, setCardsPerPage] = useState(7);
     const [currentPage, setCurrentPage] = useState(1);
-
-    const firstIndexShowedCard = cardsPerPage * (currentPage - 1);
-    const lastIndexShowedCard = firstIndexShowedCard + cardsPerPage;
-
-    const photosShowed = photos.slice(firstIndexShowedCard, lastIndexShowedCard);
     
     const pageNum = Math.ceil(photos.length / cardsPerPage);
 
     const filterImages = (newFilteredValue) => setFilteredValue(newFilteredValue);
-
-    const openModal = (id) =>  {
-        const openedCard = photos.filter(image => image.id === id)[0];
-        setIsModalOpen(true);
-        setClickedCard(openedCard);
-    };
-
-    const closeModal = () => setIsModalOpen(false);
 
     useEffect(() => {
         const getPhoto = () => {
@@ -60,13 +45,17 @@ const Photos = ({ apiUrl, album }) => {
     const getFilteredPhotos = (photos, filteredValue) => photos.filter(photoEl => photoEl.title.includes(filteredValue));
 
     const changeCurrentPage = newCurrentPage => {
+        setIsLoading(true);
         !(newCurrentPage < 1 || newCurrentPage > pageNum) 
             && setCurrentPage(newCurrentPage);
+        setIsLoading(false);
     } 
 
     const changePerPageValue = newPerPageValue => {
+        setIsLoading(true);
         setCardsPerPage(newPerPageValue);
         setCurrentPage(1);
+        setIsLoading(false);
     };
 
     return (
@@ -76,29 +65,13 @@ const Photos = ({ apiUrl, album }) => {
                 <Search filterImages={filterImages} />
             </PhotoHeadContainer>
             {isError 
-            ? <NoResult message='Problems with API. Please try again' />
-            : (isLoading 
-                ? <Loader />
-                : <Grid col='4'>
-                    {isModalOpen &&
-                        <ModalZoom 
-                            largeImg={clickedCard.url} 
-                            title={clickedCard.title}
-                            handleCloseModal={closeModal}
-                        />}
-                    {!photos.length 
-                        ? <NoResult message='No results. Please, try again' /> 
-                        : photosShowed.map(image =>
-                            <Card
-                                key={image.id} 
-                                id={image.id} 
-                                url={image.thumbnailUrl} 
-                                title={image.title} 
-                                col='4'
-                                handleZoom={openModal}
-                            />
-                        )}
-                </Grid>)}
+                ? <NoResult message='Problems with API. Please try again' />
+                : <PhotosGridWithLoading 
+                    isLoading={isLoading} 
+                    photos={photos} 
+                    cardsPerPage={cardsPerPage} 
+                    currentPage={currentPage} 
+                    />}
             {photos.length > 10 
                 && <Pagination 
                     activePage={currentPage}
